Move methods for converting VT's xml element into string to a new osdp xmlvt.py file
Because there is no other scanners, the stub methods in ospd are removed, and the wrapper
methods in daemon.py are now better located.
